+++ src/main/java/com/lulan/shincolle/entity/other/EntityFloatingFort.java
@@ -176,8 +176,8 @@ public class EntityFloatingFort extends BasicEntityAirplane {
             	    if(isTargetHurt) {
             	    	//calc kb effect
             	        if(this.kbValue > 0) {
-            	        	hitEntity.addVelocity((double)(-MathHelper.sin(rotationYaw * (float)Math.PI / 180.0F) * kbValue), 
-            	                   0.1D, (double)(MathHelper.cos(rotationYaw * (float)Math.PI / 180.0F) * kbValue));
+            	        	hitEntity.addVelocity(-MathHelper.sin(rotationYaw * (float)Math.PI / 180.0F) * kbValue, 
+            	                   0.1D, MathHelper.cos(rotationYaw * (float)Math.PI / 180.0F) * kbValue);
             	            motionX *= 0.6D;
             	            motionZ *= 0.6D;
             	        }
